requirements.txt
fastapi
pydantic
pydantic[email]
uvicorn
starlette
jinja2
SQLAlchemy
databases
bcrypt==3.2.2
passlib
python-multipart
python-dotenv
pytest
httpx
duckduckgo_search
tavily-python
langchain
langchain-google-genai
langchain-groq
langchain-huggingface
langchain-openai
langchain_community
langchain-chroma
PyMuPDF
aiofiles
arxiv
beautifulsoup4
wikipedia